requirements.txt
networkx==2.4
matplotlib==3.1.1
Django==3.0
numpy==1.16.4
djangorestframework==3.11.0
